chatbot_app/requirements.txt
slack-bolt
python-dotenv~=0.20.0
pymongo~=4.1.1
pandas~=1.4.2